+++ app/src/main/kotlin/akihz/anlaki/dev/presentation/MainActivity.kt
@@ -24,6 +24,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import akihz.anlaki.dev.data.ShizukuHelper
 import akihz.anlaki.dev.presentation.theme.AppThemeMode
 import akihz.anlaki.dev.presentation.theme.AnlakiTheme
+import akihz.anlaki.dev.utils.KeepAliveService
 import akihz.anlaki.dev.utils.PreferencesHelper
 import rikka.shizuku.Shizuku
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -71,6 +72,7 @@ class MainActivity : ComponentActivity() {
         }
 
         PreferencesHelper.init(this)
+        KeepAliveService.start(this)
 
         setContent {
             var themeMode by rememberSaveable { mutableStateOf(PreferencesHelper.themeMode) }